(Dirty Commit) Corrected visual artifact related to the creation of the swap chain
Outside of adding a depth/stencil buffer for depth testing, this still needs to be cleaned up.
I had a visual artifact of only cubes to the right of the camera failing to do any depth testing. The adjacent cube would show through the side....it was related to the multisample count given to the swap chain. It was set to 4 but needed to be 1.