Correcting memory leaks
Component storage and tear down needs to be simplified. Currently removing an entity is cumbersome and the memory management for templated components is annoying. The components might need an object factory that can handle the deallocation; or smart pointers...but I am staying old school for this (currently but plan to change it down the road).