Modified ComponentStorage class to use a vector instead of an unordered_map
While an unordered_map is N to retrieve an entity, for this super basic ECS it was easier to use a vector and more performant.  With the addition of a an object pool, cache misses would be less prevalent. Currently there isn't a substaintial performance loss with 10k simple objects. The render calls are currently expensive but I am planning to use instanced draw calls.